listen to canvas when switching renderers
